+++ Magic/CHANGELOG.md
@@ -18,6 +18,7 @@
    - Can't be spawned from command blocks, will erase any old command blocks from previous automata
    - Added some more name suffixes and prefixes
    - Increase spell attack range
+ - Add support for setting permission nodes on shop items
 
 ## 6.9.13
  - More Automata changes: